+++ src/components/Elements/CreateModal/index.ts
@@ -1,0 +1,2 @@
+export * from './CreateModal';
+export * from './CreateModal.style';
