+++ src/components/Elements/FormInputs/index.ts
@@ -1,0 +1,5 @@
+export * from './DatePicker';
+export * from './MultiLineTextInput';
+export * from './SelectInput';
+export * from './TextInput';
+export * from './UploadImage';
