--- src/components/Forms/AddAsset/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './AddAsset';
-export * from './CreateModal';
